fixed optional cast to with reflect magic
