+++ internal/value/errors.go
@@ -2,4 +2,7 @@ package value
 
 import "errors"
 
-var ErrCannotCast = errors.New("cannot cast")
+var (
+	ErrCannotCast                   = errors.New("cannot cast")
+	errDestinationTypeIsNotAPointer = errors.New("destination type is not a pointer")
+)
